+++ src/rules/Moveable.java
@@ -25,17 +25,17 @@ public abstract class Moveable {
 	
 	public GPSState movePlayer(MOVE m, SokobanState s){
 		SokobanState next = new SokobanState(s);
-		next.getBoard()[(int)s.getPlayerPos().getX()+m.getX()][(int)s.getPlayerPos().getY()+m.getY()] = next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] | TILE.PLAYER.getValue();
-		next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] = next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] & (255 - TILE.PLAYER.getValue());
+		next.getBoard()[(int)s.getPlayerPos().getX()+m.getX()][(int)s.getPlayerPos().getY()+m.getY()] = (byte)(next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] | TILE.PLAYER.getValue());
+		next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] = (byte)(next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] & (255 - TILE.PLAYER.getValue()));
 		next.movePlayer(m);
 		return next;
 	}
 	
 	public GPSState movePlayerWithBox(MOVE m, SokobanState s){
 		SokobanState next = new SokobanState(s);
-		next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()*2 ][(int)s.getPlayerPos().getY() + m.getY()*2] = next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()*2][(int)s.getPlayerPos().getY() + m.getY()*2] | TILE.BOX.getValue();
-		next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] = (next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] | TILE.PLAYER.getValue()) & (255 - TILE.BOX.getValue());
-		next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] = next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] & (255 - TILE.PLAYER.getValue());
+		next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()*2 ][(int)s.getPlayerPos().getY() + m.getY()*2] = (byte)(next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()*2][(int)s.getPlayerPos().getY() + m.getY()*2] | TILE.BOX.getValue());
+		next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] = (byte)((next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()][(int)s.getPlayerPos().getY() + m.getY()] | TILE.PLAYER.getValue()) & (255 - TILE.BOX.getValue()));
+		next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] = (byte)(next.getBoard()[(int)s.getPlayerPos().getX()][(int)s.getPlayerPos().getY()] & (255 - TILE.PLAYER.getValue()));
 		next.moveBox(m);
 		next.movePlayer(m);
 		if((next.getBoard()[(int)s.getPlayerPos().getX() + m.getX()*2 ][(int)s.getPlayerPos().getY() + m.getY()*2] & TILE.TARGET.getValue()) != 0) {
